Fix trivial signedness warnings

+++ cheevos/cheevos_client.c
@@ -1753,7 +1753,7 @@ static void rcheevos_async_award_achievement_callback(
    if (rcheevos_async_succeeded(result, &api_response.response,
             buffer, buffer_size))
    {
-      if (api_response.awarded_achievement_id != request->id)
+      if ((int)api_response.awarded_achievement_id != request->id)
          snprintf(buffer, buffer_size, "Achievement %u awarded instead",
                api_response.awarded_achievement_id);
       else if (api_response.response.error_message)
